--- BE/airbnb-be/src/main/java/com/codesquad/airbnb/accommodation/RestroomType.java
@@ -1,5 +1,0 @@
-package com.codesquad.airbnb.accommodation;
-
-public enum RestroomType {
-    PRIVATE, PUBLIC
-}
