+++ BE/airbnb-be/src/main/java/com/codesquad/airbnb/accommodation/domain/RestroomType.java
@@ -1,0 +1,5 @@
+package com.codesquad.airbnb.accommodation.domain;
+
+public enum RestroomType {
+    PRIVATE, PUBLIC
+}
